Shrink iteratively for the case when a table refers to itself as a foreign key.
Before this, if a table refers to itself with a foreign key, rows would
not be removed from the shrunk table if the foreign key row is removed.

This would lead to dumps that cannot be imported due to foreign key
violation.

This patch takes an interative approach to finding a suitable
fixed point.

https://github.com/BeautifulDestinations/postgres-subset/issues/2